created AverageStarRating.js to manage showing the average star rating for each trail. That component was added to the trail show page and each tile on the index page.
