+++ src/main/frontend/client/components/AverageStarRating.js
@@ -1,0 +1,28 @@
+import React from 'react'
+
+const StarRating = props => {
+    let starRating = ""
+    let missingStars = "★ ".repeat(5)
+    let avgRating
+
+    if (props.reviews.length > 0){
+      const allRatings = props.reviews.map(review => review.rating)
+      const totalStarRating = allRatings.reduce((total, current) => {
+        return total + current
+      })
+      avgRating = Math.round(totalStarRating/allRatings.length)
+      starRating = "★ ".repeat(avgRating)
+      missingStars = "★ ".repeat(5-avgRating)
+    } else {
+      missingStars = "Not Rated"
+    }
+
+  return (
+    <>
+      <span className="starRating">{starRating}</span>
+      <span className="missingStars">{missingStars}</span>
+    </>
+  )
+}
+
+export default StarRating
