+++ src/main/frontend/client/components/TrailTile.js
@@ -1,8 +1,10 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
+import StarRating from './StarRating'
+
 const TrailTile = props =>{
-  const {id, name, imgUrl, distance, difficulty, zipCode} = props.trail
+  const {id, name, imgUrl, distance, difficulty, zipCode, reviews} = props.trail
 
   return(
     <div className="grid-x grid-margin-x ">
@@ -13,6 +15,9 @@ const TrailTile = props =>{
       </div>
       <div className="callout secondary cell small-12 medium-6 large-4">
         <h3><Link to={`/trails/${id}`} >{name}</Link></h3>
+        <h4>
+          <StarRating reviews={reviews} />
+        </h4>
         <h4>Difficulty: {difficulty}</h4>
         <h4>Distance: {distance}</h4>
         <h4>Location: {zipCode}</h4>
